field: decompile func_800A39D8 (field sprite emitter) — 100% match
Draws one field sprite per movement accumulator: advances the accumulator a
tick, projects it with RTPS, and when the depth lands inside the ordering
table builds a POLY_FT4 billboard and links it in with addPrim. The quad's
half-extents are lerped between waypoints, scaled by FieldView::spriteScale /
OTZ, and rotated by two mvmva passes; colour and texture coords come from the
waypoint the same way.

Four things carried the match beyond the first 91.5% compile:

  * chained assignment for the u/v corner pairs (u0 = u2 = ..., v0 = v1 = ...),
    which is what makes the original load primCursor once per pair instead of
    once per store
  * separate dx/dy delta temps, with the additions as their own statements
  * s16 sprite temps — with s32 gcc keeps the two loaded bytes live across the
    divisions, where the original narrows and re-loads them (98.1% -> 99.95%)
  * building the tpage word in a variable with |= rather than one expression,
    so the OR accumulates into its first operand

Fills in what the function reveals: MoveStep 0x08..0x13 (sprite extents,
texture rect, blend mode, vertex colour), MoveRecord.zBias / .flags,
FieldSubsceneBuffer.primCursor at 0x5F20 (so the struct now spans the whole
0x5F24 field-file header), EventQueue.tpageX, and D_800C71F8 retyped from a
bare u32 to FieldView* — its first 32 bytes are the camera matrix
SetRotMatrix/SetTransMatrix already load. Adds gte_stszotz to libgte.h.

func_80041C74 / func_80040534 are the main binary's RotMatrix_gte and
TransMatrix; the field overlay links them by address so they keep func_ names.

+++ src/field/fe_object8.c
@@ -1100,7 +1100,7 @@ s32 opHandler_FACEDIRA(Eline *eline, s32 arg1) {
     if ((eline->activeMask >> eline->scriptGroup) & 1) {
         eline->turnLen = POP(eline);
         idx = POP(eline);
-        func_800A8DAC(D_80085230[idx]->field_0x256, 0x1E, D_800C71F8, buf);
+        func_800A8DAC(D_80085230[idx]->field_0x256, 0x1E, (u32)D_800C71F8, buf);
         eline->turnTgtX = D_80085230[idx]->posX / 4096;
         eline->turnTgtY = D_80085230[idx]->posY / 4096;
         eline->turnTgtZ = buf[2] + D_80085230[idx]->posZ / 4096;
@@ -1129,7 +1129,7 @@ s32 opHandler_FACEDIRP(Eline *eline, s32 arg1) {
     if ((eline->activeMask >> eline->scriptGroup) & 1) {
         eline->turnLen = POP(eline);
         slot = g_fieldVars->memberSlot[POP(eline)];
-        func_800A8DAC(slot, 0x1E, D_800C71F8, buf);
+        func_800A8DAC(slot, 0x1E, (u32)D_800C71F8, buf);
         eline->turnTgtX = D_80085224[slot].posX / 4096;
         eline->turnTgtY = D_80085224[slot].posY / 4096;
         eline->turnTgtZ = buf[2] + D_80085224[slot].posZ / 4096;
